Add new cases to DateComparisonType
New cases provide "rolling window" comparisons, making it possible to quickly check if a date falls within a specific period of time as opposed to calendar definitions of time. For example, you can now use `isInPastWeek` to check if a date falls within the past 7 days, rather than `isLastWeek` to check if a date falls in the previous calendar week.